Refactor generateSecureId to shared utility and fix encoding
